requirements.txt

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# APIs and Web
requests==2.31.0
python-dotenv==1.0.0

# Data Processing (updated versions)
pandas>=1.5.0,<2.0.0
numpy>=1.23.0,<1.25.0

# Machine Learning
scikit-learn>=1.2.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Streamlit
streamlit>=1.28.0

# OpenAI
openai>=1.3.0
